feat: Add useCalculator
